＂垃圾代码＂

+++ 安卓掌上校园服务器/src/cn/rjtraining/jdbc/FilterLogin.java
@@ -1,5 +1,6 @@
 package cn.rjtraining.jdbc;
 import java.io.IOException;
+
 import javax.servlet.Filter;
 import javax.servlet.FilterChain;
 import javax.servlet.FilterConfig;
@@ -20,7 +21,7 @@ public class FilterLogin implements Filter {
 			HttpServletResponse res = (HttpServletResponse) arg1;
 			HttpSession session = req.getSession();
 			try{
-			int uid = (Integer) session.getAttribute("uid");
+			long uid = (Long) session.getAttribute("uid");
 			String uname=(String)session.getAttribute("uname");
 			if(uid>0&&uname!=null){
 			chain.doFilter(req, res);
